iter: move Map implementation to separate files
