+++ iter/iter.go
@@ -2,11 +2,9 @@ package iter
 
 import (
 	"runtime"
-	"sync"
 	"sync/atomic"
 
 	"github.com/sourcegraph/conc"
-	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
 
 // defaultMaxGoroutines returns the default maximum number of
@@ -17,7 +15,7 @@ func defaultMaxGoroutines() int { return runtime.GOMAXPROCS(0) }
 // and ForEachIdx. The zero value is safe to use with reasonable
 // defaults.
 //
-// Iterator is safe for reuse and concurrent use.
+// Iterator is also safe for reuse and concurrent use.
 type Iterator[T any] struct {
 	// MaxGoroutines controls the maximum number of goroutines
 	// to use on this Iterator's methods.
@@ -84,34 +82,4 @@ func (iter Iterator[T]) ForEachIdx(input []T, f func(int, *T)) {
 		wg.Go(task)
 	}
 	wg.Wait()
-}
-
-// Map applies f to each element of input, returning the mapped result.
-func Map[T, R any](input []T, f func(*T) R) []R {
-	res := make([]R, len(input))
-	ForEachIdx(input, func(i int, t *T) {
-		res[i] = f(t)
-	})
-	return res
-}
-
-// MapErr applies f to each element of the input, returning the mapped result
-// and a combined error of all returned errors.
-func MapErr[T, R any](input []T, f func(*T) (R, error)) ([]R, error) {
-	var (
-		res    = make([]R, len(input))
-		errMux sync.Mutex
-		errs   error
-	)
-	ForEachIdx(input, func(i int, t *T) {
-		var err error
-		res[i], err = f(t)
-		if err != nil {
-			errMux.Lock()
-			// TODO: use stdlib errors once multierrors land in go 1.20
-			errs = errors.Append(errs, err)
-			errMux.Unlock()
-		}
-	})
-	return res, errs
 }
